Conflicts:
	src/UI/Gui.java

+++ src/Entities/BlockComparator.java
@@ -7,7 +7,7 @@ public class BlockComparator implements Comparator<Block>
 	@Override
 	public int compare(Block o1, Block o2)
 	{
-		float z = o1.getCoord().getZ() - o2.getCoord().getZ(), 
+		double z = o1.getCoord().getZ() - o2.getCoord().getZ(), 
 				y = o1.getCoord().getY() - o2.getCoord().getY(), 
 				x = o1.getCoord().getX() - o2.getCoord().getX();
